Extract the terminal command engine into terminal.js
Move the ~4.6k-line terminal subsystem out of darkmode.js into its own
module: the boot theatre, the command line and parser, the virtual filesystem
mapped onto the site nav, the ls/cat/set/pantone commands and easter eggs, the
interactive contact/subscribe flows, and the enter/exit glue (playTerminalBoot,
exitTerminal, and the flow/nav/boot/last-viewed bridge state).

The engine now drives appearance only through the window.Theme seam (aliased
once at the top of the IIFE) instead of sharing darkmode.js's closure, and it
keeps publishing the same window.Terminal / window.TerminalSlugs surfaces its
consumers (terminal-ai.js, terminal-nav.js) already depend on. darkmode.js
shrinks to the theme system it is named for; its setLayout hands control to the
terminal on entry via the new window.Terminal.enterLayout().

darkmode.js: 6924 -> 2175 lines. Wire terminal.js into head.html (after
darkmode.js, before terminal-nav/ai, in both the dev list and the prod
concat) and load it in terminal-commands.test.js. Full suite green (626).
